// supabase/functions/utils/types.ts
export enum UserStatus {
  // initial state
  none = "none",
  // activity statuses, can be live or complete, can also fall out to error
  live = "live",
  complete = "complete",
  fixed = "fixed",
  // withdrawal statuses
  funded = "funded",
  collected = "collected",
  converted = "converted",
  withdrawn = "withdrawn",
  nsf = "nsf",
  // error state
  error = "error",
}
// ...
export interface MongoSlotsUpdate {
  telegramId: string
  pointsBalance: number
  tokenBalance: number
  lastActivityAt?: Date
  energyFullNotificationTime?: Date
  slotsPlayState: SlotsPlayState
  lastLoginDay?: string
  slotsLastPlayed?: string
  loginBonusStreak?: number
}

export interface MongoUser {
  createdAt: Date
  createdDay: string
  telegramId: string
  name: string | null
  isPremium: boolean
  pointsBalance: number
  tokenBalance: number
  tonWalletAddress: string | null
  promoCode: string | null
  promoUsed: boolean
  promoUsedCount: number
  referralBonus: number
  referralTokensEarned: number
  token: string | null
  referredById: string | null
  referralCount: number
  itemsOwnedState: ItemsOwnedState
  slotsPlayState: SlotsPlayState
  dailyQuests: DailyQuests | null
  goldenSpinState: GoldenSpinState
  slotsLastPlayed: string
  lastLoginDay: string
  loginBonusStreak: number
  loginPointsToCollect: number
  lastBonusCodeTimestamp: number
  level: number
  userName: string | null
  photoUrl: string | null
  tonSpend: number | null
  starsSpend: number | null
  requestedEmail: string | null
  confirmedEmail: string | null
  lastActivityAt: Date | null
  energyFullNotificationTime: Date | null
}
// ...
export enum PackageType {
  none = "none",
  test = "test",
  levels = "levels",
  points = "points",
  tokens = "tokens",
  elixir = "elixir",
  potion = "potion",
  tool = "tool",
  treasure = "treasure",
  lootBox = "lootBox",
}

export type GoldenSpinState = {
  packageType: PackageType
  packageId: number
  result: number
  readyToCollect: boolean
  points: number
  tokens: number
}

export type SlotsResult = {
  points: number
  tokens: number
  energy: number
  level: number // you must be this level to collect the tokens listed
  treasures: number[] // treasure symbols, if any
  reels: [number, number, number]
  multipliers: [number, number, number] // multipliers for each reel
  baseScores: [number, number, number] // scores for each reel
  baseMult: number // the base multiplier applied to the score (from user level or gold potion)
  doublesMult: number // the multiplier from matching symbols on different reels
  cards?: [
    [number, number, number, number] | null,
    [number, number, number, number] | null,
    [number, number, number, number] | null,
  ] // an array of CardResult id arrays - a valid array should either be empty or have 4 entries. Null if no card symbol was spun
}

export type Achievement = {
  difficulty: "EASY" | "MEDIUM" | "HARD"
  slug: string // stringified number
  progress: number
  isFinished: boolean
  userId: string
  createdAt: Date
}

export type SlotsPlayState = {
  lastPlayed?: string
  energyAtLastPlay?: number
  result: SlotsResult
  totalPointsEarned: number
  totalTokensEarned: number
  totalSpins: number
  achievements: Achievement[]
}

export type ItemsOwnedState = {
  tool: Record<number, number>
  treasure: Record<number, number>
  elixir: Record<number, number>
  potion: Record<number, number>
}

export type DailyQuests = {
  inviteOneFriend: string // "2025-08-08"
}